requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.9.0
djoser==2.1.0
webcolors==1.11.1
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
drf-yasg==1.27.1